// ui/src/gen/ephemeralAccessAPI.ts
export interface CreateAccessRequestBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** The role template name to request. */
  roleName: string;
}

/**
 * The current access request status.
 */
export type AccessRequestResponseBodyStatus = typeof AccessRequestResponseBodyStatus[keyof typeof AccessRequestResponseBodyStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccessRequestResponseBodyStatus = {
  REQUESTED: 'REQUESTED',
  GRANTED: 'GRANTED',
  EXPIRED: 'EXPIRED',
  DENIED: 'DENIED',
  INVALID: 'INVALID',
} as const;

export interface AccessRequestResponseBody {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** The timestamp the access will expire (RFC3339 format). */
  expiresAt?: string;
  /** A human readeable description with details about the access request. */
  message?: string;
  /** The access request name. */
  name: string;
  /** The access request namespace. */
  namespace: string;
  /** The permission description of the role associated to this access request. */
  permission: string;
  /** The timestamp the access was requested (RFC3339 format). */
  requestedAt?: string;
  /** The role template associated to this access request. */
  role: string;
  /** The current access request status. */
  status?: AccessRequestResponseBodyStatus;
  /** The user associated with the access request. */
  username: string;
}